🐛 Fix captcha & add user list to groups data

+++ src/services/handlers/new-members/captcha.js
@@ -1,11 +1,18 @@
+import { Group } from '../../../entities/group.js'
 import { captchaDelay } from '../../../config/manifest.js'
 
 
 async function scheduleCaptchaInvalidation({ api, chatId, userId, greetingId }) {
-    setTimeout(async () => {
+    return setTimeout(async () => {
         try {
-            await api.deleteMessage(chatId, greetingId)
-            await api.kickChatMember(chatId, userId)
+            const userRegistered = await Group.exists({
+                groupId: chatId,
+                users: userId
+            })
+            if (!userRegistered) {
+                await api.deleteMessage(chatId, greetingId)
+                await api.kickChatMember(chatId, userId)
+            }
         } catch {
             // Never mind if I can't delete it
         }
